feat: add desktop support
Added desktop support to the project using Compose Multiplatform.
This includes adding dependencies for Linux, macOS, and Windows targets,
as well as configuring the build for desktop builds.

The `wasmJs` target was also added to enable building for the web.

Dependencies were updated to support the new targets and to use Koin for dependency injection.

Baseline profile configuration was updated to prevent automatic generation during every build.