docs: update descriptions for play store
Updated the short and full descriptions for the app on the play store.
The short description changed from "Share files between devices reliably" to "Naija Ludo: A Modern Twist on the Classic Board Game".
The full description was changed to reflect the game's features, such as its graphics, gameplay, multiplayer options, achievements, and leaderboards. Also, the full description has been re-branded to represent the app name and its developer.
The source code URL was also updated.