Enable Foojay toolchain resolver
This commit enables the Foojay toolchain resolver in `settings.gradle.kts` to automatically download and manage JDKs for the project.

It also updates the `instructions` file with new Gradle commands for running the JVM application with hot reload and the main application, and removes an outdated command. A link to an article about F-Droid has also been added.